add registry from settings
