+++ src/modules/settings/types.ts
@@ -1,3 +1,7 @@
 export type Health = {
   status: boolean;
 };
+
+export type Registries = {
+  url: string;
+};
